srm: fix srmExtendFileLifeTime implementation
Addresses several issues:

- Improves error messages

- Uses standard compliant return codes

- Fixes seconds vs milliseconds bug in SURL lifetime implementation
  (which isn't supported by dCache, so an insignificant bug).

- Cleaned up the code (lots of public fields, unused fields, cut'n'paste
  comments that didn't belong, etc).

Target: trunk
Require-notes: yes
Require-book: no
Acked-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>
Patch: http://rb.dcache.org/r/6050/

+++ modules/srm-server/src/main/java/org/dcache/srm/AbstractStorageElement.java
@@ -622,12 +622,13 @@ public interface AbstractStorageElement {
         throws SRMException;
 
     /**
+     *
      * @param newLifetime SURL lifetime in milliseconds
      *   -1 stands for infinite lifetime
      * @return long lifetime left in milliseconds
      *   -1 stands for infinite lifetime
      */
-    public int srmExtendSurlLifetime(SRMUser user, URI surl, int newLifetime)
+    public long srmExtendSurlLifetime(SRMUser user, URI surl, long newLifetime)
     throws SRMException;
 
     public String getStorageBackendVersion();
